Adición nuevas tablas

+++ Database/dbo/Tables/IngressVia.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[IngressVia]
 (
-	[IngressViaId] INT NOT NULL PRIMARY KEY IDENTITY, 
+	[IngressViaId] SMALLINT NOT NULL PRIMARY KEY IDENTITY, 
     [Code] CHAR(2) NOT NULL, 
     [Name] VARCHAR(100) NOT NULL
 )
